requirements.txt
# pip install -U -r requirements.txt

future==0.15.2
requests==2.13.0
python-dateutil==2.6.0
PyYAML==3.12
websocket==0.2.1
